requirements.txt
aiohttp>=3.13.0
streamlit>=1.32.0
pandas>=2.2.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
